requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
python-dotenv
Flask-CORS
Flask-Migrate
gunicorn
google-generativeai
